notepad requirements.txt
Flask
flask-cors
gunicorn
mysql-connector-python
Werkzeug
